chore() refactor codebase
