+++ views/not_found.erb
@@ -1,0 +1,1 @@
+<h1>  Sorry Page could not be found </h1>
